fixed PR comments next round
