+++ src/app/services/file-upload.service.ts
@@ -1,20 +1,16 @@
 import { AppConfigService } from "src/app/app-config.service";
 import { HttpClient, HttpErrorResponse, HttpHeaders } from "@angular/common/http";
-import { inject, Injectable, Injector } from "@angular/core";
+import { inject, Injectable } from "@angular/core";
 import { Observable, Subject, catchError, finalize, first, of, switchMap, tap } from "rxjs";
 import { MaybeUndefined } from "../common/app.types";
 import { LocalStorageService } from "./local-storage.service";
 import { DatasetInfo } from "../interface/navigation.interface";
 import { DatasetBasicsFragment, DatasetEndpoints } from "../api/kamu.graphql.interface";
-import { APOLLO_OPTIONS } from "apollo-angular";
 import { DatasetViewTypeEnum } from "../dataset-view/dataset-view.interface";
-import { DatasetService } from "../dataset-view/dataset.service";
 import { NavigationService } from "./navigation.service";
 import { ProtocolsService } from "./protocols.service";
 import { UploadPrepareResponse, UploadPerareData, UploadAvailableMethod } from "../common/ingest-via-file-upload.types";
-import { updateCacheHelper } from "../apollo-cache.helper";
 import { FileUploadError } from "../common/errors";
-import { takeUntilDestroyed } from "@angular/core/rxjs-interop";
 import { UnsubscribeDestroyRefAdapter } from "../common/unsubscribe.ondestroy.adapter";
 
 @Injectable({
@@ -25,8 +21,6 @@ export class FileUploadService extends UnsubscribeDestroyRefAdapter {
     private localStorageService = inject(LocalStorageService);
     private appConfigService = inject(AppConfigService);
     private navigationService = inject(NavigationService);
-    private datasetService = inject(DatasetService);
-    private injector = inject(Injector);
     private protocolsService = inject(ProtocolsService);
 
     private uploadFileLoading$ = new Subject<boolean>();
@@ -82,30 +76,11 @@ export class FileUploadService extends UnsubscribeDestroyRefAdapter {
     }
 
     private updatePage(datasetBasics: DatasetBasicsFragment): void {
-        this.updateCache(datasetBasics);
-        this.datasetService
-            .requestDatasetMainData({
-                accountName: datasetBasics.owner.accountName,
-                datasetName: datasetBasics.name,
-            })
-            .pipe(takeUntilDestroyed(this.destroyRef))
-            .subscribe(),
-            this.navigationService.navigateToDatasetView({
-                accountName: datasetBasics.owner.accountName,
-                datasetName: datasetBasics.name,
-                tab: DatasetViewTypeEnum.Overview,
-            });
-    }
-
-    private updateCache(datasetBasics: DatasetBasicsFragment): void {
-        const cache = this.injector.get(APOLLO_OPTIONS).cache;
-        if (cache) {
-            updateCacheHelper(cache, {
-                accountId: datasetBasics.owner.id,
-                datasetId: datasetBasics.id,
-                fieldNames: ["metadata"],
-            });
-        }
+        this.navigationService.navigateToDatasetView({
+            accountName: datasetBasics.owner.accountName,
+            datasetName: datasetBasics.name,
+            tab: DatasetViewTypeEnum.Overview,
+        });
     }
 
     public uploadFilePrepare(file: File): Observable<UploadPrepareResponse> {
